requirements.txt
telethon~=1.41.2
cryptg
SQLAlchemy~=2.0.44
beautifulsoup4~=4.14.2
openpyxl
aiohttp~=3.13.0
pandas~=2.3.3
pydantic~=2.11.4
tqdm~=4.67.1
python-dateutil~=2.9.0.post0
IMAPClient~=3.0.1